requirements.txt
torch
torchvision
opencv-python
Pillow
numpy
pandas
matplotlib
seaborn
scikit-learn
tqdm
joblib